Refactor map initalisation out to a module
