+++ public/src/map.js
@@ -1,0 +1,64 @@
+/* global mapboxgl, MapboxGeocoder */ // Defined by Mapbox GL JS
+
+export function initMap(apiKey) {
+  // If key wasn't set (or failed to inject) don't initalise the map section
+  if (!apiKey) {
+    document.getElementById("map").innerHTML = "Map appears here. API key not configured (see contributing guidelines).";
+    return;
+  }
+
+  const long = -3.1883;
+  const lat = 55.9533;
+  const coor = [long, lat];
+  mapboxgl.accessToken = apiKey;
+  const map = new mapboxgl.Map({
+    container: 'map', // ID in the HTML
+    style: 'mapbox://styles/mapbox/light-v10',
+    center: coor,
+    zoom: 12,
+  });
+
+  const geocoder = new MapboxGeocoder({
+    accessToken: apiKey,
+    mapboxgl: mapboxgl,
+    placeholder: 'Search for places in United Kingdom',
+    bbox: [-8.196671, 50.064075, 1.737475, 60.917070],
+    proximity: {
+      longitude: long,
+      latitude: lat
+    },
+    marker: false
+  });
+
+  // Add the geocoder to the map
+  map.addControl(geocoder);
+
+  // After the map style has loaded on the page,
+  // add a source layer and default styling for a single point
+  map.on('load', function () {
+    map.addSource('single-point', {
+      type: 'geojson',
+      data: {
+        type: 'FeatureCollection',
+        features: []
+      }
+    });
+
+    map.addLayer({
+      id: 'point',
+      source: 'single-point',
+      type: 'circle',
+      paint: {
+        'circle-radius': 10,
+        'circle-color': '#448ee4'
+      }
+    });
+
+    // Listen for the `result` event from the Geocoder
+    // `result` event is triggered when a user makes a selection
+    //  Add a marker at the result's coordinates
+    geocoder.on('result', function (e) {
+      map.getSource('single-point').setData(e.result.geometry);
+    });
+  });
+}
